Prevent deleting a product if it still has a ticket assigned to it

+++ app/Http/Controllers/ProductController.php
@@ -224,6 +224,11 @@ class ProductController extends Controller
             return Redirect::back();
         }
 
+        if ($product->ticket) {
+            Session::flash('flash_message', 'You cannot delete this product is a ticket associated with it.');
+            return Redirect::back();
+        }
+
         $product->delete();
 
         Session::flash('flash_message', 'The product has been deleted.');
